Move PlaylistsResultsScreen to correct namespace
