+++ osu.Game/Screens/Multi/Playlists/PlaylistsPlayer.cs
@@ -14,7 +14,6 @@ using osu.Game.Online.API;
 using osu.Game.Online.Rooms;
 using osu.Game.Rulesets;
 using osu.Game.Scoring;
-using osu.Game.Screens.Multi.Ranking;
 using osu.Game.Screens.Play;
 using osu.Game.Screens.Ranking;
